+++ infrastructure/src/main/java/app/majime/infrastructure/lims/RestConstants.java
@@ -1,0 +1,22 @@
+package app.majime.infrastructure.lims;
+
+public class RestConstants {
+
+    public static final String APPLICATION_NAME = "/majime";
+    public static final String API_VERSION_1 = "/v1";
+
+    public static final String RESOURCE_MANUFACTURER = "/manufacturer";
+    public static final String RESOURCE_SAMPLE = "/sample";
+    public static final String RESOURCE_USER = "/user";
+    public static final String RESOURCE_MATERIAL = "/material";
+    public static final String RESOURCE_OUTOFSPEC = "/outofspec";
+    public static final String RESOURCE_PARAMETER = "/parameter";
+    public static final String RESOURCE_RESULT = "/result";
+    public static final String RESOURCE_UNIT = "/unit";
+    public static final String RESOURCE_BATCH = "/batch";
+    public static final String RESOURCE_SUPPLIER = "/supplier";
+    public static final String RESOURCE_CERTIFICATE = "/certificate";
+    public static final String RESOURCE_ADDRESS = "/address";
+    public static final String RESOURCE_LAB = "/lab";
+    public static final String RESOURCE_SPECIFICATION = "/specification";
+}
