--- infrastructure/src/main/java/app/majime/infrastructure/lims/batch/repositories/BatchRepository.java
@@ -1,10 +1,0 @@
-package app.majime.infrastructure.lims.batch.repositories;
-
-import app.majime.core.batch.Batch;
-import org.springframework.data.repository.CrudRepository;
-
-import java.util.Optional;
-
-public interface BatchRepository extends CrudRepository<Batch,Long> {
-    Optional<Batch> findById(Long id);
-}
