--- infrastructure/src/main/java/app/majime/infrastructure/lims/manufacturer/repositories/ManufacturerRepository.java
@@ -1,8 +1,0 @@
-package app.majime.infrastructure.lims.manufacturer.repositories;
-import app.majime.core.manufacturer.Manufacturer;
-import org.springframework.data.repository.CrudRepository;
-import java.util.Optional;
-
-public interface ManufacturerRepository extends CrudRepository<Manufacturer,Long> {
-    Optional<Manufacturer> findByName(String name);
-}
