--- infrastructure/src/main/java/app/majime/infrastructure/lims/material/repositories/MaterialRepository.java
@@ -1,13 +1,0 @@
-package app.majime.infrastructure.lims.material.repositories;
-
-import app.majime.core.material.Material;
-import org.springframework.data.repository.CrudRepository;
-
-import java.util.Optional;
-
-public interface MaterialRepository extends CrudRepository<Material,Long>{
-    Optional<Material> findByName(String name);
-
-}
-
-
