+++ infrastructure/src/main/java/app/majime/infrastructure/lims/specification/SpecificationRepository.java
@@ -1,0 +1,7 @@
+package app.majime.infrastructure.lims.specification;
+
+import app.majime.core.specification.Specification;
+import org.springframework.data.repository.CrudRepository;
+
+public interface SpecificationRepository extends CrudRepository<Specification, Long> {
+}
