--- infrastructure/src/main/java/app/majime/infrastructure/lims/supplier/repositories/SupplierRepository.java
@@ -1,10 +1,0 @@
-package app.majime.infrastructure.lims.supplier.repositories;
-
-import app.majime.core.supplier.Supplier;
-import org.springframework.data.repository.CrudRepository;
-
-import java.util.Optional;
-
-public interface SupplierRepository extends CrudRepository<Supplier,Long> {
-    Optional<Supplier> findByName(String name);
-}
