+++ infrastructure/src/main/java/app/majime/infrastructure/lims/unit/UnitRepository.java
@@ -1,0 +1,10 @@
+package app.majime.infrastructure.lims.unit;
+
+import app.majime.core.unit.Unit;
+import org.springframework.data.repository.CrudRepository;
+
+import java.util.Optional;
+
+public interface UnitRepository extends CrudRepository<Unit,Long> {
+    Optional<Unit> findByName(String name);
+}
