--- infrastructure/src/main/java/app/majime/infrastructure/lims/user/repositories/UserRepository.java
@@ -1,11 +1,0 @@
-package app.majime.infrastructure.lims.user.repositories;
-
-import app.majime.core.sample.Sample;
-import app.majime.core.user.User;
-import org.springframework.data.repository.CrudRepository;
-
-import java.util.Optional;
-
-public interface UserRepository extends CrudRepository<User,Long> {
-    Optional<User> findByEmail(String email);
-}
